fix some minor warnings

+++ progs/xdemos/glxheads.c
@@ -1,4 +1,4 @@
-/* $Id: glxheads.c,v 1.2 2000/11/10 17:23:07 brianp Exp $ */
+/* $Id: glxheads.c,v 1.3 2002/03/08 19:44:28 brianp Exp $ */
 
 /*
  * Exercise multiple GLX connections on multiple X displays.
@@ -30,6 +30,7 @@
 #include <GL/glx.h>
 #include <stdio.h>
 #include <stdlib.h>
+#include <string.h>
 #include <unistd.h>
 
 
@@ -247,9 +248,9 @@ static void
 PrintInfo(const struct head *h)
 {
    printf("Name: %s\n", h->DisplayName);
-   printf("  Display:     0x%x\n", h->Dpy);
-   printf("  Window:      0x%x\n", h->Win);
-   printf("  Context:     0x%x\n", h->Context);
+   printf("  Display:     0x%x\n", (int) h->Dpy);
+   printf("  Window:      0x%x\n", (int) h->Win);
+   printf("  Context:     0x%x\n", (int) h->Context);
    printf("  GL_VERSION:  %s\n", h->Version);
    printf("  GL_VENDOR:   %s\n", h->Vendor);
    printf("  GL_RENDERER: %s\n", h->Renderer);
